# openapi/voicelayerd.v1.yaml
openapi: 3.1.0
info:
  title: VoiceLayer Local Control API
  version: 0.2.0
  description: |
    Local control API exposed by the VoiceLayer daemon over a Unix domain socket.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost
    description: Logical placeholder. The daemon is served over a Unix domain socket.
paths:
...
  /v1/providers:
    get:
      operationId: listProviders
      summary: List supported providers and host adapters.
      responses:
        "200":
          description: Provider descriptors.
          content:
            application/json:
              schema:
                type: object
                required: [providers]
                properties:
                  providers:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProviderDescriptor"
        "503":
          $ref: "#/components/responses/Problem"
...
  /v1/sessions/dictation:
    post:
      operationId: startLiveDictationSession
      summary: Create a live dictation session and start recording audio.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartDictationRequest"
      responses:
        "200":
          description: Dictation session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureSession"
        "400":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
  /v1/sessions/dictation/stop:
    post:
      operationId: stopLiveDictationSession
      summary: Stop a live dictation session, transcribe the recorded audio, and return the result.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StopDictationRequest"
      responses:
        "200":
          description: Dictation capture result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DictationCaptureResult"
        "404":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
        "502":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
        "504":
          $ref: "#/components/responses/Problem"
  /v1/sessions/compose:
    post:
      operationId: createCompositionJob
      summary: Register a composition request and return a preview artifact.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComposeRequest"
      responses:
        "200":
          description: Composition job receipt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompositionReceipt"
        "400":
          $ref: "#/components/responses/Problem"
        "502":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /v1/rewrites:
    post:
      operationId: rewriteText
      summary: Register a rewrite request and return a preview artifact.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewriteRequest"
      responses:
        "200":
          description: Rewrite job receipt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompositionReceipt"
        "400":
          $ref: "#/components/responses/Problem"
        "502":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /v1/translations:
    post:
      operationId: translateText
      summary: Register a translation request and return a preview artifact.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranslateRequest"
      responses:
        "200":
          description: Translation job receipt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompositionReceipt"
        "400":
          $ref: "#/components/responses/Problem"
        "502":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /v1/transcriptions:
    post:
      operationId: transcribeAudioFile
      summary: Transcribe a local audio file through the configured ASR provider.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscribeRequest"
      responses:
        "200":
          description: Transcription result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionResult"
        "400":
          $ref: "#/components/responses/Problem"
        "502":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /v1/dictation/capture:
    post:
      operationId: captureDictation
      summary: Record a short clip, transcribe it, and return the resulting dictation session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DictationCaptureRequest"
      responses:
        "200":
          description: Dictation capture result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DictationCaptureResult"
        "400":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
        "502":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /v1/inject:
    post:
      operationId: planInjection
      summary: Build an injection plan for a target. Pure function; always 200.
      description: Computes the injection payload for the target (bracketed paste, etc.) locally.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InjectRequest"
      responses:
        "200":
          description: Injection plan response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InjectionPlan"
components:
  responses:
    Problem:
      description: RFC 9457 problem details describing the failure.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          description: Problem type URN, e.g. `urn:voicelayer:problem:provider_unavailable`.
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
      description: |
        Problem type registry:
        - `urn:voicelayer:problem:bad_request` (400)
        - `urn:voicelayer:problem:session_not_found` (404)
        - `urn:voicelayer:problem:recording_failed` (500)
        - `urn:voicelayer:problem:internal` (500)
        - `urn:voicelayer:problem:provider_request_failed` (502)
        - `urn:voicelayer:problem:provider_unavailable` (503)
        - `urn:voicelayer:problem:worker_unavailable` (503)
        - `urn:voicelayer:problem:worker_timeout` (504)
...
    ProviderDescriptor:
      type: object
      required:
        - id
        - kind
        - transport
        - local
        - default_enabled
        - experimental
        - license
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [asr, llm, tts, host_adapter]
        transport:
          type: string
        local:
          type: boolean
        default_enabled:
          type: boolean
        experimental:
          type: boolean
        license:
          type: string
    LanguageProfile:
      type: object
      required: [strategy, input_languages]
      properties:
        strategy:
          type: string
          enum: [auto_detect, locked, follow_previous, bilingual]
        input_languages:
          type: array
          items:
            type: string
        output_language:
          type:
            - string
            - "null"
    StartDictationRequest:
      type: object
      required: [trigger]
      properties:
        trigger:
          type: string
          enum: [push_to_talk, toggle, cli, tui, tray_button]
        language_profile:
          $ref: "#/components/schemas/LanguageProfile"
        translate_to_english:
          type: boolean
        keep_audio:
          type: boolean
        segmentation:
          $ref: "#/components/schemas/SegmentationMode"
        provider_id:
          description: |
            Optional ASR provider for every transcription this dictation
            session emits. When omitted the daemon routes to the
            configured whisper.cpp chain (whisper-server preferred,
            whisper-cli fallback). Selecting `mimo_v2_5_asr` opts the
            entire session into the Xiaomi MiMo-V2.5-ASR backend, and
            `qwen3_asr_1_7b` opts into the Alibaba Qwen3-ASR-1.7B
            backend; see `docs/guides/local-asr-provider.html` for the
            hardware preconditions and the opt-in environment variables
            for each. The value is shared by every `TranscribeRequest`
            the daemon issues for this session — fixed/vad-gated
            segments, the stop-time transcription, and the one-shot
            path. An unrecognized id surfaces as `provider unavailable`
            rather than falling back to whisper.
          type:
            - string
            - "null"
          enum:
            - whisper_cpp
            - mimo_v2_5_asr
            - qwen3_asr_1_7b
            - null
    SegmentationMode:
      description: |
        Controls how the recorder subprocess is rolled during a live dictation
        session. When omitted the daemon uses `one_shot`, which keeps the
        pre-Phase-3 behavior (a single recording across the whole session).
        `fixed` rotates the recorder every `segment_secs` seconds so each
        chunk can be transcribed while the next segment captures. `vad_gated`
        rotates at the shorter `probe_secs` cadence and runs a silero-vad
        classification on every probe; speech probes accumulate into a
        pending buffer that flushes to transcribe on the first run of
        `silence_gap_probes` silent probes or when the buffered duration
        reaches `max_segment_secs` (whichever comes first).
      oneOf:
        - type: object
          required: [mode]
          properties:
            mode:
              type: string
              enum: [one_shot]
        - type: object
          required: [mode, segment_secs]
          properties:
            mode:
              type: string
              enum: [fixed]
            segment_secs:
              type: integer
              minimum: 1
              description: Duration of each captured segment in whole seconds.
        - type: object
          required: [mode, probe_secs, max_segment_secs]
          properties:
            mode:
              type: string
              enum: [vad_gated]
            probe_secs:
              type: integer
              minimum: 1
              description: |
                Duration of each classification probe in whole seconds. Short
                values (1-2 s) keep flush latency close to the speaker's
                natural pauses; the orchestrator still pays one
                `segment_probe` RPC per probe, so very small values trade
                responsiveness for RPC overhead.
            max_segment_secs:
              type: integer
              minimum: 1
              description: |
                Upper bound on the buffered speech span before a forced
                flush, in whole seconds. Prevents an unbroken utterance from
                accumulating indefinitely when the speaker never pauses.
            silence_gap_probes:
              type: integer
              minimum: 0
              default: 1
              description: |
                Number of consecutive silent probes required to flush the
                pending speech buffer. `1` (the default) flushes on the
                first silent probe after speech; higher values tolerate
                brief breathing gaps inside a single speech unit.
    StopDictationRequest:
      type: object
      required: [session_id]
      properties:
        session_id:
          type: string
          format: uuid
    DictationCaptureRequest:
      type: object
      required: [trigger, duration_seconds]
      properties:
        trigger:
          type: string
          enum: [push_to_talk, toggle, cli, tui, tray_button]
        language_profile:
          $ref: "#/components/schemas/LanguageProfile"
        duration_seconds:
          type: integer
          minimum: 1
        translate_to_english:
          type: boolean
        keep_audio:
          type: boolean
        provider_id:
          description: |
            Optional ASR provider for the single transcription this
            capture emits. Defaults to the configured whisper.cpp chain
            when omitted; `mimo_v2_5_asr` opts into the Xiaomi
            MiMo-V2.5-ASR backend, and `qwen3_asr_1_7b` opts into the
            Alibaba Qwen3-ASR-1.7B backend. See
            `TranscribeRequest.provider_id` for the unrecognized-id
            semantics.
          type:
            - string
            - "null"
          enum:
            - whisper_cpp
            - mimo_v2_5_asr
            - qwen3_asr_1_7b
            - null
    CaptureSession:
      type: object
      required:
        - session_id
        - mode
        - state
        - trigger
        - language_profile
        - created_at_millis
      properties:
        session_id:
          type: string
          format: uuid
        mode:
          type: string
          enum: [dictation, compose, rewrite, translate]
        state:
          type: string
          enum: [idle, listening, transcribing, previewing, awaiting_confirmation, completed, failed]
        trigger:
          type: string
          enum: [push_to_talk, toggle, cli, tui, tray_button]
        language_profile:
          $ref: "#/components/schemas/LanguageProfile"
        created_at_millis:
          type: integer
          format: int64
    ComposeRequest:
      type: object
      required: [spoken_prompt]
      properties:
        spoken_prompt:
          type: string
        archetype:
          type: string
          enum:
            - email
            - cover_letter
            - daily_report
            - issue
            - pull_request_description
            - prompt
            - technical_summary
            - custom
        output_language:
          type:
            - string
            - "null"
    RewriteRequest:
      type: object
      required: [source_text, style]
      properties:
        source_text:
          type: string
        style:
          type: string
          enum: [more_formal, shorter, politer, more_technical, translate]
        output_language:
          type:
            - string
            - "null"
    TranslateRequest:
      type: object
      required: [source_text, target_language]
      properties:
        source_text:
          type: string
        target_language:
          type: string
    TranscribeRequest:
      type: object
      required: [audio_file]
      properties:
        audio_file:
          type: string
        language:
          type:
            - string
            - "null"
        translate_to_english:
          type: boolean
        provider_id:
          description: |
            Optional ASR provider identifier. Defaults to the configured
            whisper.cpp chain (whisper-server preferred, whisper-cli
            fallback). The optional `mimo_v2_5_asr` provider routes to
            the Xiaomi MiMo-V2.5-ASR backend, and `qwen3_asr_1_7b`
            routes to the Alibaba Qwen3-ASR-1.7B backend, both through
            the Python worker; see `docs/guides/local-asr-provider.html`
            for the hardware preconditions and the opt-in environment
            variables. An unrecognized provider id surfaces as a
            `provider unavailable` error rather than falling back to
            whisper.
          type:
            - string
            - "null"
          enum:
            - whisper_cpp
            - mimo_v2_5_asr
            - qwen3_asr_1_7b
            - null
    PreviewArtifact:
      type: object
      required: [artifact_id, status, title, notes]
      properties:
        artifact_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [ready, needs_provider, rejected]
        title:
          type: string
        generated_text:
          type:
            - string
            - "null"
        notes:
          type: array
          items:
            type: string
    CompositionReceipt:
      type: object
      required: [job_id, preview]
      properties:
        job_id:
          type: string
          format: uuid
        preview:
          $ref: "#/components/schemas/PreviewArtifact"
    InjectRequest:
      type: object
      required: [target, text]
      properties:
        target:
          type: string
          enum:
            - gui_accessible
            - gui_clipboard
            - terminal_bracketed_paste
            - terminal_kitty_remote
        text:
          type: string
        auto_submit:
          type: boolean
          default: false
    InjectionPlan:
      type: object
      required: [target, payload, auto_submit]
      properties:
        target:
          type: string
          enum:
            - gui_accessible
            - gui_clipboard
            - terminal_bracketed_paste
            - terminal_kitty_remote
        payload:
          type: string
        auto_submit:
          type: boolean
    TranscriptionResult:
      type: object
      required: [text, notes]
      properties:
        text:
          type: string
        detected_language:
          type:
            - string
            - "null"
        notes:
          type: array
          items:
            type: string
    DictationCaptureResult:
      type: object
      required: [session, transcription]
      properties:
        session:
          $ref: "#/components/schemas/CaptureSession"
        transcription:
          $ref: "#/components/schemas/TranscriptionResult"
        audio_file:
          type:
            - string
            - "null"